MatrixPilot_trunk: very minor changes to bring trunk into line with my development branch. DPRINT some feedback in LedTest program, cleanup of I2C1 driver, and introduce some stardardising of types in altitude module.
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@3324 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ libDCM/estAltitude.c
@@ -37,10 +37,10 @@ int barometer_temperature_gnd = 0;
 long barometer_altitude;        // above sea level altitude - ASL (millimeters)
 long barometer_agl_altitude;    // above ground level altitude - AGL
 long barometer_pressure;
-int barometer_temperature;
+int16_t barometer_temperature;
 float sea_level_pressure;
 
-inline int get_barometer_temperature(void)   { return barometer_temperature; }
+inline int16_t get_barometer_temperature(void)   { return barometer_temperature; }
 inline long get_barometer_pressure(void)     { return barometer_pressure; }
 inline long get_barometer_altitude(void)     { return barometer_altitude; }
 inline long get_barometer_agl_altitude(void) { return barometer_agl_altitude; }
